chatbot-backend/requirements.txt
llama-index==0.14.3
chromadb==1.1.0
fastapi==0.115.0
uvicorn[standard]==0.37.0
beautifulsoup4==4.14.2
requests==2.32.5
python-dotenv==1.1.1
pydantic==2.11.9
pydantic-settings==2.11.0
openai==1.109.1
aiofiles==24.1.0
lxml==6.0.2